Extracted ExpansionConverter
